Deprecates passing `$control` to `Orchestra\Html\Form\Field::getField()`.
Technically `$control` is an instanceof `Field` which make it rather pointless and confusing to pass `$this` as a parameter. Therefore, the parameter is removed in 3.3.x.

However for 3.1 and 3.2, we can't just remove the parameter as this would create a huge breaking backward compatibility, so instead we would convert `$control` to `$templates` if
`$control` is an array and `$templates` is empty.

Signed-off-by: crynobone <crynobone@gmail.com>